+++ src/plonemeeting/restapi/services/config.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+CONFIG_ID_ERROR = "The \"config_id\" parameter must be given!"
+CONFIG_ID_NOT_FOUND_ERROR = "The given \"config_id\" named \"%s\" was not found!"